adicionado usecase para processamento de arquivos com transacoes

+++ internal/application/dto/product_dto.go
@@ -3,20 +3,20 @@ package dto
 type CreateProductInputDto struct {
 	Title        string  `json:"title"`
 	ProducerName string  `json:"producer_id"`
-	Value        float64 `json:"value"`
+	Value        int `json:"value"`
 }
 
 type CreateProductOutputDto struct {
 	Title        string  `json:"title"`
 	ProducerName string  `json:"producer_name"`
-	Value        float64 `json:"value"`
+	Value        int `json:"value"`
 }
 
 type ProductOutputDto struct {
 	ID           string  `json:"id"`
 	Title        string  `json:"title"`
 	ProducerName string  `json:"producer_name"`
-	Value        float64 `json:"value"`
+	Value        int `json:"value"`
 }
 
 type ListProductsOutputDto struct {
